Fix tests for moving of events and properties back to the lifecycle (removal from public api).
