+++ test/unit.js
@@ -1,10 +1,8 @@
 import './boot';
 import './unit/api/create';
 import './unit/api/emit';
-import './unit/api/events';
 import './unit/api/init';
 import './unit/api/no-conflict';
-import './unit/api/properties';
 import './unit/attributes';
 import './unit/constructor';
 import './unit/document-observer';
@@ -14,6 +12,8 @@ import './unit/ignoring';
 import './unit/lifecycle';
 import './unit/lifecycle-scenarios';
 import './unit/lifecycle/created';
+import './unit/lifecycle/events';
+import './unit/lifecycle/properties';
 import './unit/registration';
 import './unit/registry';
 import './unit/templating';
